requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==1.1.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
drf-extensions==0.3.1
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.0
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jwt==0.5.4
Markdown==3.0.1
MarkupSafe==1.1.0
oauthlib==2.0.7
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pycparser==2.18
pycryptodome==3.6.1
PyJWT==1.6.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
